+++ src/render/opengl/opengldevice.h
@@ -13,7 +13,7 @@ public:
 	virtual std::unique_ptr<IGPUBuffer> createGPUBuffer(size_t size, uint32_t usage) override;
 	virtual std::unique_ptr<IRenderPass> createRenderPass(SRenderPassParams&) override;
 	virtual std::unique_ptr<IPipeline> createPipeline(SPipelineParams& params) override;
-	virtual std::unique_ptr<ITexture> createTexture(ITexture::EFormat format, uint32_t usage, uint16_t width, uint16_t height, bool bMipmapped) override;
+	virtual std::unique_ptr<ITexture> createTexture(ITexture::EFormat format, uint32_t usage, uint32_t width, uint32_t height, bool bMipmapped) override;
 	virtual void finishJobs() override;
 
 	PFNGLCLEARNAMEDFRAMEBUFFERFVPROC glClearNamedFramebufferfv;
